Moving the domain name contraint into the field namespace
