+++ src/Entity/Validation/Constraints/FieldConstraints/DomainName.php
@@ -1,6 +1,6 @@
 <?php declare(strict_types=1);
 
-namespace EdmondsCommerce\DoctrineStaticMeta\Entity\Validation\Constraints;
+namespace EdmondsCommerce\DoctrineStaticMeta\Entity\Validation\Constraints\FieldConstraints;
 
 use Symfony\Component\Validator\Constraint;
